+++ src/elements/element-x-axis.ts
@@ -10,36 +10,8 @@ import {
   timeYear,
 } from "d3-time";
 
-import { Colors } from "../helpers";
+import { Colors, multiFormat } from "../helpers";
 import { getNumXTicks } from "../helpers/helpers-axis";
-import { timeFormat } from "d3-time-format";
-
-const formatMillisecond = timeFormat(".%L");
-const formatSecond = timeFormat(":%S");
-const formatMinute = timeFormat("%H:%M");
-const formatHour = timeFormat("%H:%M");
-const formatDay = timeFormat("%a %d");
-const formatWeek = timeFormat("%b %d");
-const formatMonth = timeFormat("%B");
-const formatYear = timeFormat("%Y");
-
-function multiFormat(date: Date) {
-  return (timeSecond(date) < date
-    ? formatMillisecond
-    : timeMinute(date) < date
-    ? formatSecond
-    : timeHour(date) < date
-    ? formatMinute
-    : timeDay(date) < date
-    ? formatHour
-    : timeMonth(date) < date
-    ? timeWeek(date) < date
-      ? formatDay
-      : formatWeek
-    : timeYear(date) < date
-    ? formatMonth
-    : formatYear)(date);
-}
 
 function addXAxisPath(
   ctx: CanvasRenderingContext2D,
@@ -47,8 +19,6 @@ function addXAxisPath(
   pixelRatio: number
 ) {
   ctx.strokeStyle = "#fff";
-
-  const tickFormat = multiFormat;
 
   const xRange = xScale.range();
   const numXTicks = getNumXTicks(xRange[1] - xRange[0]);
@@ -60,7 +30,7 @@ function addXAxisPath(
     ctx.textBaseline = "top";
     ctx.textAlign = "center";
     ctx.font = `${TICK_LABEL_FONT_SIZE}px ${TICK_LABEL_FONT}`;
-    ctx.fillText(tickFormat(tick), xScale(tick), 9);
+    ctx.fillText(multiFormat(tick), xScale(tick), 9);
     ctx.closePath();
   });
 }
